+++ src/main/java/com/quyet/superapp/controller/ReportController.java
@@ -1,6 +1,7 @@
 package com.quyet.superapp.controller;
 
 import com.quyet.superapp.entity.Report;
+import com.quyet.superapp.service.ReportService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -10,6 +11,7 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/reports")
 @RequiredArgsConstructor
+
 public class ReportController {
 
     private final ReportService service;
@@ -20,7 +22,7 @@ public class ReportController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Report> getById(@PathVariable Integer id) {
+    public ResponseEntity<Report> getById(@PathVariable Long id) {
         return service.getById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -32,13 +34,14 @@ public class ReportController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Report> update(@PathVariable Integer id, @RequestBody Report report) {
+    public ResponseEntity<Report> update(@PathVariable Long id, @RequestBody Report report) {
         Report updated = service.update(id, report);
         return updated != null ? ResponseEntity.ok(updated) : ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Integer id) {
+    public void delete(@PathVariable Long id) {
         service.delete(id);
     }
 }
+
