+++ src/main/java/com/quyet/superapp/service/DonationRegistrationService.java
@@ -17,7 +17,7 @@ public class DonationRegistrationService {
         return repository.findAll();
     }
 
-    public Optional<DonationRegistration> getById(Integer id) {
+    public Optional<DonationRegistration> getById(Long id) {
         return repository.findById(id);
     }
 
@@ -25,7 +25,7 @@ public class DonationRegistrationService {
         return repository.save(registration);
     }
 
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         repository.deleteById(id);
     }
 }
